refactor: mundado id de Integer para Long fix: requisiçao rest

+++ src/main/java/com/vitormarcal/asteya/divida/DividaRepository.java
@@ -4,8 +4,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface DividaRepository extends JpaRepository<Divida, Integer>{
+public interface DividaRepository extends JpaRepository<Divida, Long>{
 
-    Optional<Divida> findById(Integer idDivida);
+    Optional<Divida> findById(Long idDivida);
 
 }
